moved out of focus to a module
